feat: lesson 03_hacky_hello_world

+++ src/bsp/raspberrypi/cpu.zig
@@ -1,3 +1,3 @@
 const std = @import("std");
 
-pub export const BOOT_CORE_ID: u8 linksection(".text.start_arguments") = 0;
+pub export const BOOT_CORE_ID: usize linksection(".text.start_arguments") = 0;
